Move betty.plugins.event_type to betty.event_types
