+++ betty/event_types/burial.py
@@ -1,5 +1,5 @@
 """
-The funeral event type.
+The burial event type.
 """
 
 from __future__ import annotations
@@ -7,20 +7,20 @@ from __future__ import annotations
 from typing import final
 
 from betty.event_type import EventType, EventTypeDefinition
+from betty.event_types.death import Death
 from betty.locale.localizable.gettext import _, ngettext
-from betty.plugins.event_type.death import Death
 
 
 @final
 @EventTypeDefinition(
-    "funeral",
-    label=_("Funeral"),
-    label_plural=_("Funerals"),
-    label_countable=ngettext("{count} funeral", "{count} funerals"),
+    "burial",
+    label=_("Burial"),
+    label_plural=_("Burials"),
+    label_countable=ngettext("{count} burial", "{count} burials"),
     after={Death},
     indicates=Death,
 )
-class Funeral(EventType):
+class Burial(EventType):
     """
-    .. plugin:: event-type:funeral.
+    .. plugin:: event-type:burial.
     """
