+++ betty/event_types/cremation.py
@@ -7,8 +7,8 @@ from __future__ import annotations
 from typing import final
 
 from betty.event_type import EventType, EventTypeDefinition
+from betty.event_types.death import Death
 from betty.locale.localizable.gettext import _, ngettext
-from betty.plugins.event_type.death import Death
 
 
 @final
