+++ betty/event_types/divorce_announcement.py
@@ -7,11 +7,11 @@ from __future__ import annotations
 from typing import final
 
 from betty.event_type import EventType, EventTypeDefinition
+from betty.event_types.birth import Birth
+from betty.event_types.death import Death
+from betty.event_types.divorce import Divorce
+from betty.event_types.marriage import Marriage
 from betty.locale.localizable.gettext import _, ngettext
-from betty.plugins.event_type.birth import Birth
-from betty.plugins.event_type.death import Death
-from betty.plugins.event_type.divorce import Divorce
-from betty.plugins.event_type.marriage import Marriage
 
 
 @final
