+++ betty/event_types/retirement.py
@@ -7,9 +7,9 @@ from __future__ import annotations
 from typing import final
 
 from betty.event_type import EventType, EventTypeDefinition
+from betty.event_types.birth import Birth
+from betty.event_types.death import Death
 from betty.locale.localizable.gettext import _, ngettext
-from betty.plugins.event_type.birth import Birth
-from betty.plugins.event_type.death import Death
 
 
 @final
