+++ betty/plugins/extension/_theme/__init__.py
@@ -11,12 +11,12 @@ from typing import TYPE_CHECKING, cast
 
 from betty.date import AnyDate, Date
 from betty.entity import persistent_id
+from betty.event_types.birth import Birth
+from betty.event_types.death import Death
 from betty.functools import unique
 from betty.plugins.entity.event import Event
 from betty.plugins.entity.person import Person
 from betty.plugins.entity.place import Place
-from betty.plugins.event_type.birth import Birth
-from betty.plugins.event_type.death import Death
 from betty.privacy.resolve import is_public
 from betty.roles.subject import Subject
