+++ betty/tests/event_types/test_death.py
@@ -3,11 +3,11 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from betty.date import Date
+from betty.event_types.birth import Birth
+from betty.event_types.death import Death
 from betty.plugins.entity.event import Event
 from betty.plugins.entity.person import Person
 from betty.plugins.entity.presence import Presence
-from betty.plugins.event_type.birth import Birth
-from betty.plugins.event_type.death import Death
 from betty.roles.subject import Subject
 
 if TYPE_CHECKING:
