+++ betty/tests/plugins/asset/raspberry_mint/templates/entity/test_event_dimensions.py
@@ -1,12 +1,12 @@
 from betty.date import Date
 from betty.document import Document, EntityContexts
+from betty.event_types.birth import Birth
 from betty.plugins.asset_directory.raspberry_mint import RASPBERRY_MINT
 from betty.plugins.entity.citation import Citation
 from betty.plugins.entity.event import Event
 from betty.plugins.entity.place import Place
 from betty.plugins.entity.place_name import PlaceName
 from betty.plugins.entity.source import Source
-from betty.plugins.event_type.birth import Birth
 from betty.test_utils.conftest import AssertTemplateFile
