+++ betty/tests/plugins/asset/raspberry_mint/templates/entity/test_label__event.py
@@ -1,10 +1,10 @@
 from betty.document import Document, EntityContexts
+from betty.event_types.birth import Birth
+from betty.event_types.marriage import Marriage
 from betty.plugins.asset_directory.raspberry_mint import RASPBERRY_MINT
 from betty.plugins.entity.event import Event
 from betty.plugins.entity.person import Person
 from betty.plugins.entity.presence import Presence
-from betty.plugins.event_type.birth import Birth
-from betty.plugins.event_type.marriage import Marriage
 from betty.roles.subject import Subject
 from betty.test_utils.conftest import AssertTemplateFile
