+++ betty/tests/plugins/enricher/deriver/test___init__.py
@@ -3,14 +3,14 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from betty.date import Date, DateRange
+from betty.event_types.birth import Birth
+from betty.event_types.death import Death
+from betty.event_types.residence import Residence
 from betty.load import load
 from betty.plugins.enricher.deriver import Deriver
 from betty.plugins.entity.event import Event
 from betty.plugins.entity.person import Person
 from betty.plugins.entity.presence import Presence
-from betty.plugins.event_type.birth import Birth
-from betty.plugins.event_type.death import Death
-from betty.plugins.event_type.residence import Residence
 from betty.roles.subject import Subject
 from betty.test_utils.entity import record_added
