+++ betty/tests/plugins/entity/test_person.py
@@ -6,6 +6,7 @@ import pytest
 
 from betty.entity import Entity
 from betty.entity.association import AssociationRequired, TemporaryToOneResolver
+from betty.event_types.birth import Birth
 from betty.genders.non_binary import NonBinary
 from betty.genders.unknown import Unknown as UnknownGender
 from betty.locale import DEFAULT_LOCALE, DEFAULT_LOCALE_TAG, to_language_tag
@@ -16,7 +17,6 @@ from betty.plugins.entity.person import Person
 from betty.plugins.entity.person_name import PersonName
 from betty.plugins.entity.presence import Presence
 from betty.plugins.entity.source import Source
-from betty.plugins.event_type.birth import Birth
 from betty.privacy import Privacy
 from betty.roles.subject import Subject
 from betty.test_utils.entity import EntityTestBase
