+++ betty/tests/plugins/entity/test_place.py
@@ -7,13 +7,13 @@ from geopy import Point
 
 from betty.entity import Entity
 from betty.entity.association import AssociationRequired, TemporaryToOneResolver
+from betty.event_types.birth import Birth
 from betty.locale import DEFAULT_LOCALE, DEFAULT_LOCALE_TAG, to_language_tag
 from betty.plugins.entity.enclosure import Enclosure
 from betty.plugins.entity.event import Event
 from betty.plugins.entity.link import Link
 from betty.plugins.entity.place import Place
 from betty.plugins.entity.place_name import PlaceName
-from betty.plugins.event_type.birth import Birth
 from betty.plugins.place_type.hamlet import Hamlet
 from betty.plugins.place_type.unknown import Unknown as UnknownPlaceType
 from betty.test_utils.entity import EntityTestBase
