+++ betty/tests/privacy/test_privatizer.py
@@ -6,6 +6,9 @@ from typing import TYPE_CHECKING
 import pytest
 
 from betty.date import Date, DateRange
+from betty.event_types.birth import Birth
+from betty.event_types.death import Death
+from betty.event_types.marriage import Marriage
 from betty.plugins.entity.citation import Citation
 from betty.plugins.entity.enclosure import Enclosure
 from betty.plugins.entity.event import Event
@@ -15,9 +18,6 @@ from betty.plugins.entity.person import Person
 from betty.plugins.entity.place import Place
 from betty.plugins.entity.presence import Presence
 from betty.plugins.entity.source import Source
-from betty.plugins.event_type.birth import Birth
-from betty.plugins.event_type.death import Death
-from betty.plugins.event_type.marriage import Marriage
 from betty.privacy import Privacy
 from betty.privacy.privatizer import Privatizer
 from betty.project import DEFAULT_LIFETIME_THRESHOLD
