+++ pyproject.toml
@@ -187,32 +187,32 @@ betty = 'betty.console:main_standalone'
 'source' = 'betty.plugins.entity.source:Source'
 
 [project.entry-points.'betty.event_type']
-'adoption' = 'betty.plugins.event_type.adoption:Adoption'
-'baptism' = 'betty.plugins.event_type.baptism:Baptism'
-'bar-mitzvah' = 'betty.plugins.event_type.bar_mitzvah:BarMitzvah'
-'bat-mitzvah' = 'betty.plugins.event_type.bat_mitzvah:BatMitzvah'
-'birth' = 'betty.plugins.event_type.birth:Birth'
-'burial' = 'betty.plugins.event_type.burial:Burial'
-'conference' = 'betty.plugins.event_type.conference:Conference'
-'confirmation' = 'betty.plugins.event_type.confirmation:Confirmation'
-'correspondence' = 'betty.plugins.event_type.correspondence:Correspondence'
-'cremation' = 'betty.plugins.event_type.cremation:Cremation'
-'death' = 'betty.plugins.event_type.death:Death'
-'divorce' = 'betty.plugins.event_type.divorce:Divorce'
-'divorce-announcement' = 'betty.plugins.event_type.divorce_announcement:DivorceAnnouncement'
-'emigration' = 'betty.plugins.event_type.emigration:Emigration'
-'engagement' = 'betty.plugins.event_type.engagement:Engagement'
-'funeral' = 'betty.plugins.event_type.funeral:Funeral'
-'immigration' = 'betty.plugins.event_type.immigration:Immigration'
-'marriage' = 'betty.plugins.event_type.marriage:Marriage'
-'marriage-announcement' = 'betty.plugins.event_type.marriage_announcement:MarriageAnnouncement'
-'missing' = 'betty.plugins.event_type.missing:Missing'
-'occupation' = 'betty.plugins.event_type.occupation:Occupation'
+'adoption' = 'betty.event_types.adoption:Adoption'
+'baptism' = 'betty.event_types.baptism:Baptism'
+'bar-mitzvah' = 'betty.event_types.bar_mitzvah:BarMitzvah'
+'bat-mitzvah' = 'betty.event_types.bat_mitzvah:BatMitzvah'
+'birth' = 'betty.event_types.birth:Birth'
+'burial' = 'betty.event_types.burial:Burial'
+'conference' = 'betty.event_types.conference:Conference'
+'confirmation' = 'betty.event_types.confirmation:Confirmation'
+'correspondence' = 'betty.event_types.correspondence:Correspondence'
+'cremation' = 'betty.event_types.cremation:Cremation'
+'death' = 'betty.event_types.death:Death'
+'divorce' = 'betty.event_types.divorce:Divorce'
+'divorce-announcement' = 'betty.event_types.divorce_announcement:DivorceAnnouncement'
+'emigration' = 'betty.event_types.emigration:Emigration'
+'engagement' = 'betty.event_types.engagement:Engagement'
+'funeral' = 'betty.event_types.funeral:Funeral'
+'immigration' = 'betty.event_types.immigration:Immigration'
+'marriage' = 'betty.event_types.marriage:Marriage'
+'marriage-announcement' = 'betty.event_types.marriage_announcement:MarriageAnnouncement'
+'missing' = 'betty.event_types.missing:Missing'
+'occupation' = 'betty.event_types.occupation:Occupation'
 'project-configuration' = 'betty.project._plugin_discovery:_discover_event_types'
-'residence' = 'betty.plugins.event_type.residence:Residence'
-'retirement' = 'betty.plugins.event_type.retirement:Retirement'
-'unknown' = 'betty.plugins.event_type.unknown:Unknown'
-'will' = 'betty.plugins.event_type.will:Will'
+'residence' = 'betty.event_types.residence:Residence'
+'retirement' = 'betty.event_types.retirement:Retirement'
+'unknown' = 'betty.event_types.unknown:Unknown'
+'will' = 'betty.event_types.will:Will'
 
 [project.entry-points.'betty.extension']
 'http-api-doc' = 'betty.plugins.extension.http_api_doc:HttpApiDoc'
